requirements.txt
clip-retrieval
scikit-image
git+https://github.com/openai/CLIP.git
torch
Pillow
pytesseract